Move registry only endpoints from api to registry
policies, profiles, envs, and services were declared in the api package,
and did not yet exist in the registry package. Since their endpoints are
all registry only, move them over directly.
